requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
chromedriver-binary==95.0.4638.17.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
dj-database-url==0.5.0
Django==3.2.8
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
lxml==4.6.3
Markdown==3.3.4
psycopg2==2.9.1
pytz==2021.3
requests==2.26.0
selenium==3.141.0
soupsieve==2.2.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
webdriver-manager==3.5.1
whitenoise==5.3.0
zipp==3.6.0
